+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [3.1.0] - Unreleased
 
+### Added
+- WordPress events list and viewer display component's logo if it is known.
+
 ### Changed
 - There's now a `decalog_no_log_post_activity` filter to modify ability to log posts activities.